doc: remove test-npm from general build doc
`make test-npm` is not particularly robust (currently fails on Linux and
Windows, reportedly) and results in a fair number of requests for help
from people new to the project. It is used when upgrading npm in core,
which only a small number of mostly-predefined people do. Remove it from
the general build doc.

Refs: https://github.com/nodejs/node/issues/12836
PR-URL: https://github.com/nodejs/node/pull/12840
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Jeremiah Senkpiel <fishrock123@rocketmail.com>
Reviewed-By: Daniel Bevenius <daniel.bevenius@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>